Data type changed

+++ FactorialProgram.java
@@ -4,8 +4,8 @@ public class FactorialProgram {
 
 	public static void main(String[] args) {
 		// Factorial Program--Changes
-		int factorialNumber=5;
-		long factorialValue=1;
+		Float factorialNumber=5;
+		Double factorialValue=1;
 		for(int i=factNumber;i>=1;i--) {
 			factValue=factValue*i;
 		}
